requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
defusedxml==0.8.0rc2
Django==4.2
django-allauth==65.1.0
django-filter==24.3
django-oauth-toolkit==3.0.1
djangorestframework==3.15.2
drf-social-oauth2==3.1.0
environs==11.0.0
idna==3.10
jwcrypto==1.5.6
marshmallow==3.23.0
oauthlib==3.2.2
packaging==24.1
pycparser==2.22
PyJWT==2.9.0
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
